refactor: enforce package domain boundaries
