+++ src/codex_usage_tracker/diagnostics/snapshot_events.py
@@ -5,14 +5,20 @@ from __future__ import annotations
 import hashlib
 import json
 import re
-import shlex
 from collections import Counter
 from pathlib import Path
 from typing import Any
 
-SAFE_LABEL_RE = re.compile(r"^[A-Za-z0-9_.:-]{1,80}$")
+from codex_usage_tracker.core.command_parsing import (
+    GIT_COMMAND_ROOTS,
+    SENSITIVE_LABEL_PREFIXES,
+    command_tokens,
+    git_operation,
+    looks_like_assignment,
+    strip_command_wrappers,
+)
+
 SAFE_PATH_LABEL_RE = re.compile(r"^[A-Za-z0-9_.:@*+-]{1,80}$")
-SENSITIVE_LABEL_PREFIXES = ("sk-", "sk_", "ghp_", "github_pat_", "xox")
 SHELL_TOOL_NAMES = {
     "bash",
     "exec_command",
@@ -22,12 +28,8 @@ SHELL_TOOL_NAMES = {
     "terminal",
     "write_stdin",
 }
-SHELL_EXEC_WRAPPERS = {"bash", "fish", "sh", "zsh"}
-SHELL_EXEC_COMMAND_OPTIONS = {"-c", "-lc", "-cl"}
-RUN_WRAPPERS = {"npx", "pipenv", "poetry", "uv"}
 READ_COMMAND_ROOTS = {"cat", "find", "grep", "head", "nl", "rg", "sed", "strings", "tail", "wc"}
 SEARCH_READ_ROOTS = {"find", "rg"}
-GIT_COMMAND_ROOTS = {"git", "gh"}
 GIT_OPERATION_CATEGORIES = {
     "add": ("local_mutation", "local_mutation"),
     "branch": ("read_only", "read_only"),
@@ -67,21 +69,6 @@ GH_OPERATION_CATEGORIES = {
     "repo": ("repository", "github_remote"),
     "run": ("workflow", "github_remote"),
     "workflow": ("workflow", "github_remote"),
-}
-GIT_GLOBAL_OPTIONS_WITH_VALUES = {
-    "-C",
-    "-c",
-    "--config-env",
-    "--git-dir",
-    "--namespace",
-    "--super-prefix",
-    "--work-tree",
-}
-GH_GLOBAL_OPTIONS_WITH_VALUES = {
-    "-H",
-    "-R",
-    "--hostname",
-    "--repo",
 }
 READ_PRODUCTIVITY_NOTE = (
     "Read-to-modify counts are temporal correlations: a read is counted when the same "
@@ -131,18 +118,10 @@ def is_shell_tool(function_name: str) -> bool:
     return lowered in SHELL_TOOL_NAMES or suffix in SHELL_TOOL_NAMES
 
 
-def command_root_and_child(command: str) -> tuple[str, str]:
-    tokens = _strip_command_wrappers(_command_tokens(command))
-    if not tokens:
-        return "unknown_command", "unknown"
-    root = _command_root(tokens)
-    return root, _command_child(root, tokens)
-
-
 def read_path_refs_from_command(command: str, *, root: str) -> list[dict[str, str]]:
     if root not in READ_COMMAND_ROOTS:
         return []
-    tokens = _strip_command_wrappers(_command_tokens(command))
+    tokens = strip_command_wrappers(command_tokens(command))
     if not tokens:
         return []
     path_tokens = _read_path_tokens(root=root, tokens=tokens)
@@ -160,10 +139,10 @@ def read_path_refs_from_command(command: str, *, root: str) -> list[dict[str, st
 def git_interaction_from_command(command: str, *, root: str) -> dict[str, str] | None:
     if root not in GIT_COMMAND_ROOTS:
         return None
-    tokens = _strip_command_wrappers(_command_tokens(command))
+    tokens = strip_command_wrappers(command_tokens(command))
     if not tokens:
         return None
-    operation = _git_operation(tokens, root=root)
+    operation = git_operation(tokens, root=root)
     if operation is None:
         return None
     if root == "git":
@@ -245,18 +224,6 @@ def original_output_count(output: object) -> int | None:
 
 def optional_str(value: object) -> str | None:
     return value if isinstance(value, str) and value else None
-
-
-def safe_label(value: object) -> str | None:
-    if not isinstance(value, str):
-        return None
-    stripped = value.strip()
-    lowered = stripped.lower()
-    if lowered.startswith(SENSITIVE_LABEL_PREFIXES):
-        return None
-    if "/" in stripped or "\\" in stripped:
-        return None
-    return lowered if SAFE_LABEL_RE.fullmatch(stripped) else None
 
 
 def simple_rows(
@@ -427,7 +394,7 @@ def _path_values(value: object) -> list[str]:
 
 def _path_ref_from_token(token: str) -> dict[str, str] | None:
     raw = token.strip()
-    if not raw or raw == "-" or _is_shell_separator(raw) or _looks_like_assignment(raw):
+    if not raw or raw == "-" or _is_shell_separator(raw) or looks_like_assignment(raw):
         return None
     if raw.startswith(("$", "`")) or "://" in raw:
         return None
@@ -453,136 +420,5 @@ def _stable_hash(value: str) -> str:
     return hashlib.sha256(value.encode("utf-8")).hexdigest()[:12]
 
 
-def _command_tokens(command: str) -> list[str]:
-    try:
-        return shlex.split(command, posix=True)
-    except ValueError:
-        return []
-
-
-def _strip_command_wrappers(tokens: list[str]) -> list[str]:
-    remaining = list(tokens)
-    while remaining:
-        while remaining and _looks_like_assignment(remaining[0]):
-            remaining.pop(0)
-        if not remaining:
-            break
-        base = _basename(remaining[0])
-        if base in {"command", "env", "sudo"}:
-            remaining.pop(0)
-            continue
-        if base in SHELL_EXEC_WRAPPERS:
-            nested = _shell_exec_nested_tokens(remaining)
-            if nested:
-                remaining = nested
-                continue
-        if base in RUN_WRAPPERS and len(remaining) > 2 and remaining[1] == "run":
-            remaining = remaining[2:]
-            continue
-        if base == "npx" and len(remaining) > 1:
-            remaining = remaining[1:]
-            continue
-        break
-    return remaining
-
-
-def _command_root(tokens: list[str]) -> str:
-    base = _basename(tokens[0])
-    if base in {"py.test", "pytest"}:
-        return "pytest"
-    module_root = _python_module_root(tokens)
-    if module_root:
-        return module_root
-    if base == "py" or base == "python" or base.startswith("python"):
-        return "python"
-    return safe_label(base) or "unknown_command"
-
-
-def _command_child(root: str, tokens: list[str]) -> str:
-    module_child = _python_module_child(root, tokens)
-    if module_child is not None:
-        return module_child
-    if root == "python":
-        for index, token in enumerate(tokens[:-1]):
-            if token == "-m":
-                module = safe_label(_basename(tokens[index + 1]).split(".", 1)[0])
-                return f"-m:{module}" if module else "-m:unknown"
-        return tokens[1] if len(tokens) > 1 and tokens[1].startswith("-") else "<script>"
-    if root in GIT_COMMAND_ROOTS:
-        return _git_operation(tokens, root=root) or "<none>"
-    if len(tokens) <= 1:
-        return "<none>"
-    child = safe_label(_basename(tokens[1]))
-    return child or "<arg>"
-
-
-def _git_operation(tokens: list[str], *, root: str) -> str | None:
-    option_args = GIT_GLOBAL_OPTIONS_WITH_VALUES if root == "git" else GH_GLOBAL_OPTIONS_WITH_VALUES
-    skip_next = False
-    for token in tokens[1:]:
-        if skip_next:
-            skip_next = False
-            continue
-        if _is_shell_separator(token):
-            break
-        if token == "--":
-            continue
-        if token.startswith("-"):
-            option_name = token.split("=", 1)[0]
-            if option_name in option_args and "=" not in token:
-                skip_next = True
-            continue
-        return safe_label(_basename(token))
-    return "<none>"
-
-
-def _python_module_child(root: str, tokens: list[str]) -> str | None:
-    if root not in {"mypy", "pytest", "ruff"}:
-        return None
-    for index, token in enumerate(tokens[:-1]):
-        if token != "-m":
-            continue
-        if index + 2 < len(tokens):
-            child = tokens[index + 2]
-            if child.startswith("-"):
-                return child
-            return "<target>"
-        return "<none>"
-    return None
-
-
 def _is_shell_separator(token: str) -> bool:
     return token in {"&&", "||", ";", "|"}
-
-
-def _looks_like_assignment(token: str) -> bool:
-    return bool(re.fullmatch(r"[A-Za-z_][A-Za-z0-9_]*=.*", token))
-
-
-def _shell_exec_nested_tokens(tokens: list[str]) -> list[str] | None:
-    for index, token in enumerate(tokens[1:], start=1):
-        if token not in SHELL_EXEC_COMMAND_OPTIONS:
-            continue
-        if index + 1 >= len(tokens):
-            return None
-        nested = _command_tokens(tokens[index + 1])
-        return nested or [tokens[index + 1]]
-    return None
-
-
-def _python_module_root(tokens: list[str]) -> str | None:
-    base = _basename(tokens[0]) if tokens else ""
-    if not (base == "py" or base == "python" or base.startswith("python")):
-        return None
-    for index, token in enumerate(tokens[:-1]):
-        if token != "-m":
-            continue
-        module = safe_label(_basename(tokens[index + 1]).split(".", 1)[0])
-        if module in {"mypy", "pytest", "ruff"}:
-            return module
-        return None
-    return None
-
-
-def _basename(token: str) -> str:
-    return re.split(r"[\\/]", token)[-1].lower().lstrip("$")
