+++ src/codex_usage_tracker/diagnostics/snapshot_source_scan.py
@@ -6,16 +6,15 @@ from collections import Counter
 from pathlib import Path
 from typing import Any
 
+from codex_usage_tracker.core.command_parsing import command_root_and_child, safe_label
 from codex_usage_tracker.diagnostics.snapshot_events import (
     allocate_token_count,
-    command_root_and_child,
     git_interaction_from_command,
     is_shell_tool,
     optional_str,
     original_output_count,
     read_path_refs_from_command,
     read_reader,
-    safe_label,
     shell_command_from_payload,
     simple_rows,
     unique_path_rows,
