+++ src/codex_usage_tracker/reports/tach.domain.toml
@@ -1,7 +1,8 @@
 [root]
 depends_on = [
-    "//codex_usage_tracker.core",
-    "//codex_usage_tracker.diagnostics",
-    "//codex_usage_tracker.pricing",
-    "//codex_usage_tracker.store",
+  "//codex_usage_tracker.allowance_intelligence",
+  "//codex_usage_tracker.core",
+  "//codex_usage_tracker.diagnostics",
+  "//codex_usage_tracker.pricing",
+  "//codex_usage_tracker.store",
 ]
