+++ tests/diagnostics/test_diagnostic_snapshot_events.py
@@ -2,12 +2,7 @@ from __future__ import annotations
 
 import json
 
-import pytest
-
-from codex_usage_tracker.diagnostics.snapshot_events import (
-    command_root_and_child,
-    shell_command_from_payload,
-)
+from codex_usage_tracker.diagnostics.snapshot_events import shell_command_from_payload
 
 
 def test_shell_command_from_payload_reads_supported_argument_shapes() -> None:
@@ -34,25 +29,3 @@ def test_shell_command_from_payload_reads_supported_argument_shapes() -> None:
         == "fallback"
     )
     assert shell_command_from_payload({"cmd": "git status"}, function_name="read_file") is None
-
-
-@pytest.mark.parametrize(
-    ("command", "expected_root"),
-    [
-        ("bash -lc 'rg TODO src'", "rg"),
-        ("zsh -lc 'git status --short'", "git"),
-        ("sh -c 'nl -ba src/app.py'", "nl"),
-        ("python -m pytest -q", "pytest"),
-        ("python3 -m mypy", "mypy"),
-        ("uv run pytest -q", "pytest"),
-        ("poetry run python -m pytest", "pytest"),
-        ("npx eslint .", "eslint"),
-        ("npm run test", "npm"),
-        ("$PWCLI status", "pwcli"),
-    ],
-)
-def test_command_root_and_child_normalizes_common_wrappers(
-    command: str, expected_root: str
-) -> None:
-    root, _child = command_root_and_child(command)
-    assert root == expected_root
